Adds harpy (large) wings (#1393)
This time everything is good, holy fuck, last one was weird. Now I know
how to do includes.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

I've grown tired of not being able to make an avali of my dreams. I am
not a fan of teshari sprites and it's just generally not what I was
going for. So I present to you... **_Dramatic drum rolls_** THE Harpy
(large) WINGS!
The sprites have been stolen ported from Splurt.
https://github.com/WoolyAypa is the artist that made them, or at least
they have made a commit with those sprites... pretty sure.
btw this is the second time I'm making this PR because the last branch I
was working on.... broke a lil, but this one's all fine and dandy.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

Wings, WINGS, **WINGS!!** Who doesn't like having more wing options!
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

![image](https://github.com/Bubberstation/Bubberstation/assets/51271713/a6b2ad83-a1fd-4859-aade-6b40868ef2ae)

![image](https://github.com/Bubberstation/Bubberstation/assets/51271713/458eba00-e495-4e49-ae79-784bbaa89023)
Tested them, everything's good
<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Added Harpy (large) wings
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->